Change customerID to long type
Causes less type conversion, so doesn't seem like customerID should be String

+++ group5.ics372.pa1/group5/ics372/pa1/Company.java
@@ -18,7 +18,7 @@ import java.util.List;
  */
 public class Company {
 
-	private int nextCustomerId;
+	private long nextCustomerId;
 	private Catalog catalog;
 	private CustomerList customerList;
 	private long nextApplianceId;
